requirements.txt
flask==3.0.3
firebase-admin==6.5.0
google-generativeai==0.7.2
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.0.3
